refactor(trainer-clients): consolidate client detail into three sub-tabs
개요 becomes the always-visible detail header (alert badges + today's
metrics + the two actions), and 루틴 folds into 운동, so 채팅/식단/운동
are the only tabs and the prescription sits next to its execution.
Extracts the alert pill duplicated between the dashboard and the header
into a shared AlertBadge.

Review caught the workout tab gating its whole list on the history
provider, which would have hidden the routines whenever /history failed;
each section now owns its async state, with a regression test.

flutter analyze clean; 321 tests pass.
