+++ frontend/flutter_trainer/test/features/clients/client_detail_workout_test.dart
@@ -1,4 +1,5 @@
 import 'package:drift/native.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:oncare_trainer/features/dashboard/domain/dashboard_summary.dart'
@@ -6,6 +7,7 @@ import 'package:oncare_trainer/features/dashboard/domain/dashboard_summary.dart'
 import 'package:oncare_trainer/app/router/routes.dart';
 import 'package:oncare_trainer/core/storage/app_database.dart';
 import 'package:oncare_trainer/core/storage/seed_data.dart';
+import 'package:oncare_trainer/features/clients/domain/entities/routine_history_entry.dart';
 import 'package:oncare_trainer/shared/services/client_repository.dart';
 
 import '../../helpers/pump_app.dart';
@@ -16,6 +18,15 @@ const Map<String, String> seedClientIds = <String, String>{
   '이지수': 'seed-client-2',
   '박성호': 'seed-client-3',
 };
+
+/// Fails only `watchHistory`; every other read still succeeds.
+class _HistoryFailsRepository extends DriftClientRepository {
+  const _HistoryFailsRepository(super.db);
+
+  @override
+  Stream<List<RoutineHistoryEntry>> watchHistory(String clientId) =>
+      Stream<List<RoutineHistoryEntry>>.error(Exception('history down'));
+}
 
 void main() {
   group('ClientRepository.watchHistory', () {
@@ -75,14 +86,21 @@ void main() {
       final expected = (counted.reduce((a, b) => a + b) / counted.length)
           .round();
 
+      // The tab now opens on the routines it absorbed from the old 루틴
+      // tab, so the completion card starts below the fold.
+      expect(find.text('배정된 루틴'), findsOneWidget);
+      await tester.scrollUntilVisible(find.text('이번 주 완료율'), 150);
       expect(find.text('이번 주 완료율'), findsOneWidget);
       expect(find.text('$expected%'), findsOneWidget);
-      expect(find.text('완료'), findsOneWidget);
+      // '완료' is also a PT session's status in the card above, so the
+      // legend is matched loosely; 부분/미완료 are legend-only.
+      expect(find.text('완료'), findsWidgets);
       expect(find.text('부분'), findsOneWidget);
       expect(find.text('미완료'), findsOneWidget);
 
       // History entries with feedback + note boxes. Lower list items are
       // built lazily — scroll each into view before asserting.
+      await tester.scrollUntilVisible(find.text('7/12 (오늘)'), 150);
       expect(find.text('7/12 (오늘)'), findsOneWidget);
       expect(find.text('100%'), findsWidgets);
       await tester.scrollUntilVisible(find.text('트레이너 메모'), 150);
@@ -92,6 +110,34 @@ void main() {
       // A skipped exercise line renders (struck-through content present).
       await tester.scrollUntilVisible(find.text('스트레칭 (생략)'), 150);
       expect(find.text('스트레칭 (생략)'), findsOneWidget);
+    });
+
+    testWidgets('a failed 운동 기록 load does not take the routines with it', (
+      tester,
+    ) async {
+      // 배정된 루틴 and the PT sessions used to be their own tab, so they
+      // stayed reachable when the history endpoint was down. Gating the
+      // whole list on the history provider silently undid that.
+      await pumpTrainerApp(
+        tester,
+        token: 'demo-trainer-token',
+        at: AppRoutes.clientDetail(seedClientIds['김민수']!, section: 'workout'),
+        extraOverrides: <Override>[
+          clientRepositoryProvider.overrideWith(
+            (ref) => _HistoryFailsRepository(ref.watch(appDatabaseProvider)),
+          ),
+        ],
+      );
+
+      // The sections above the history still rendered — they are the
+      // first thing on the tab, so the failure did not blank it.
+      expect(find.text('배정된 루틴'), findsOneWidget);
+      expect(find.text('PT 프로그램 이력'), findsOneWidget);
+      await tester.scrollUntilVisible(find.text('이번 주 완료율'), 150);
+      expect(find.text('이번 주 완료율'), findsOneWidget);
+      // The failure is reported in place, where the history would be.
+      await tester.scrollUntilVisible(find.text('운동 기록을 불러오지 못했어요'), 150);
+      expect(find.text('운동 기록을 불러오지 못했어요'), findsOneWidget);
     });
   });
 }
